Move Metrics to separate class
