+++ src/main/java/org/opentripplanner/standalone/configure/ConstructApplication.java
@@ -1,6 +1,5 @@
 package org.opentripplanner.standalone.configure;
 
-import io.micrometer.core.instrument.Metrics;
 import javax.annotation.Nullable;
 import javax.ws.rs.core.Application;
 import org.opentripplanner.datastore.api.DataSource;
